MAINT: Remove superfluous "whitelist" utility for ingredients
This could always have been implemented with just the source
