+++ core/src/zeit/wochenmarkt/interfaces.py
@@ -16,22 +16,6 @@ class IRecipeCategoriesSource(zope.schema.interfaces.IIterableSource):
     """Available categories."""
 
 
-class IIngredientsWhitelist(zope.interface.Interface):
-    """The whitelist contains all selectable ingredients providing
-    `IIngredient`.
-    """
-
-    def search(term):
-        """Return a list of ingredients whose names contain the given term."""
-
-    def get(id):
-        """Return the ingredient for the given id."""
-
-
-class IIngredientsSource(zope.schema.interfaces.IIterableSource):
-    """Available ingredients."""
-
-
 class IIngredient(zope.interface.Interface):
     """An ingredient item in a list of IIngredients as part of
     IIngredientSource.
@@ -40,3 +24,7 @@ class IIngredient(zope.interface.Interface):
     code = zope.schema.TextLine(title='Internal ingredient id')
 
     name = zope.schema.TextLine(title='User visible name of ingredient')
+
+
+class IIngredientsSource(zope.schema.interfaces.IIterableSource):
+    """Available ingredients."""
