+++ core/src/zeit/wochenmarkt/tests/test_ingredients.py
@@ -1,7 +1,5 @@
 # -*- coding: utf-8 -*-
 
-import zope.component
-
 import zeit.cms.testing
 import zeit.wochenmarkt.interfaces
 import zeit.wochenmarkt.testing
@@ -11,26 +9,21 @@ class TestIngredients(zeit.wochenmarkt.testing.FunctionalTestCase):
     """Testing ..ingredients.Ingredients"""
 
     def test_ingredient_should_be_found_through_xml(self):
-        ingredients = zope.component.getUtility(
-            zeit.wochenmarkt.interfaces.IIngredientsWhitelist
-        )._load()
-        basmati = dict(ingredients.items()).get('basmatireis')
+        source = zeit.wochenmarkt.sources.ingredientsSource(None)
+        basmati = source.find('basmatireis')
         assert 'Basmatireis' == basmati.name
         assert ['Reis', 'Basmati'] == basmati.qwords
         assert 'Basmatireis' == basmati.singular
         assert 'Basmatireis' == basmati.plural
 
     def test_ingredient_should_be_found_by_id(self):
-        calamari = zope.component.getUtility(zeit.wochenmarkt.interfaces.IIngredientsWhitelist).get(
-            'calamari'
-        )
+        source = zeit.wochenmarkt.sources.ingredientsSource(None)
+        calamari = source.find('calamari')
         assert 'Calamari' == calamari.name
         assert ['Tintenfisch', 'Kalamar'] == calamari.qwords
 
     def test_autocomplete_should_be_available_for_ingredients(self):
-        result = zope.component.getUtility(
-            zeit.wochenmarkt.interfaces.IIngredientsWhitelist
-        ).search('B')
+        result = zeit.wochenmarkt.sources.ingredientsSource(None).factory.search('B')
         assert 8 == len(result)
         names = []
         for item in result:
@@ -38,9 +31,7 @@ class TestIngredients(zeit.wochenmarkt.testing.FunctionalTestCase):
         assert 'Brathähnchen' in names
 
     def test_ingredients_should_be_sorted_with_exact_match_leading(self):
-        result = zope.component.getUtility(
-            zeit.wochenmarkt.interfaces.IIngredientsWhitelist
-        ).search('ei')
+        result = zeit.wochenmarkt.sources.ingredientsSource(None).factory.search('ei')
         assert ['Ei', 'Eis', 'Basmatireis', 'Brei'] == ([r.name for r in result])
